verify(): Support skipping else block if its value is Never.
Basically, if a particular if (testCondition) testCondition has
the type Never, in either the ifBranch or elseBranch.

Then we know that either the `if` or `else` block will never
run given the type of the testCondition expression.

This allows one to skip dead if/else branches in functions
which have overloaded types. For example a function may have the
type:

```
fn(err: Error, value: undefined);
fn(err: null, value: String);
```

We can have an `if (err)` expr and correctly know that the value
type will always be a string in the `else` block because when
the `err` value is `Error` the else block test condition has the
value `Never`, aka statically, the else block will never be
evaluated. Which means it's safe to treat the second argument
`value` as a string, even though it's possible undefined.